third_party/blink/web_tests/virtual/select-option-accessibility-target-size-disabled/fast/forms/select-popup/popup-menu-accessibility-minimum-target-size-expected.txt

This is a testharness.js-based test.
[FAIL] The minimum height of the option element in the popup menu should be 24 CSS pixels.
  assert_greater_than_equal: The minimum height of the option element in the popup menu should be 24 CSS pixels. expected a number greater than or equal to 24 but got 17
Harness: the test ran to completion.
